spec: Don't try to rebuild webpack when building released tarballs
Release tarballs already have dist/. We can't rebuild the webpack in an
RPM as that requires npm and network access for `npm install`. But we
also don't want to ship the entire `node_modules/` as that is huge and
would require lots of careful license review. So ship a stub
node_modules/ that satisfies the Makefile dependency.

Closes #36